# clients/openapi-v2.json
{
 "components": {
  "schemas": {
...
   "AssembleRequest": {
    "properties": {
     "preset_ids": {
      "items": {
       "type": "string"
      },
      "title": "Preset Ids",
      "type": "array"
     },
     "profile_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Profile Id"
     },
     "storyboard_id": {
      "title": "Storyboard Id",
      "type": "string"
     },
     "title": {
      "default": "",
      "maxLength": 120,
      "title": "Title",
      "type": "string"
     }
    },
    "required": [
     "storyboard_id"
    ],
    "title": "AssembleRequest",
    "type": "object"
   },
   "Assembled": {
    "properties": {
     "episode": {
      "$ref": "#/components/schemas/Episode"
     },
     "series": {
      "$ref": "#/components/schemas/Series"
     }
    },
    "required": [
     "series",
     "episode"
    ],
    "title": "Assembled",
    "type": "object"
   },
...
   "Bible-Output": {
    "additionalProperties": false,
    "properties": {
     "characters": {
      "items": {
       "$ref": "#/components/schemas/Character-Output"
      },
      "title": "Characters",
      "type": "array"
     },
     "locations": {
      "items": {
       "$ref": "#/components/schemas/Location-Output"
      },
      "title": "Locations",
      "type": "array"
     },
     "props": {
      "items": {
       "$ref": "#/components/schemas/Prop-Output"
      },
      "title": "Props",
      "type": "array"
     },
     "styles": {
      "items": {
       "$ref": "#/components/schemas/Style-Output"
      },
      "title": "Styles",
      "type": "array"
     }
    },
    "required": [
     "characters",
     "locations",
     "props",
     "styles"
    ],
    "title": "Bible",
    "type": "object"
   },
...
   "Character-Output": {
    "additionalProperties": false,
    "properties": {
     "age": {
      "description": "Mio only generates adult characters.",
      "minimum": 18.0,
      "title": "Age",
      "type": "integer"
     },
     "appearance": {
      "description": "Checklist used by VLM QA.",
      "items": {
       "type": "string"
      },
      "title": "Appearance",
      "type": "array"
     },
     "description": {
      "default": "",
      "title": "Description",
      "type": "string"
     },
     "gender": {
      "$ref": "#/components/schemas/Gender",
      "default": "unknown"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "loras": {
      "items": {
       "$ref": "#/components/schemas/LoraRef-Output"
      },
      "title": "Loras",
      "type": "array"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "outfits": {
      "additionalProperties": {
       "items": {
        "type": "string"
       },
       "type": "array"
      },
      "title": "Outfits",
      "type": "object"
     },
     "references": {
      "items": {
       "$ref": "#/components/schemas/AssetRef-Output"
      },
      "title": "References",
      "type": "array"
     },
     "signature": {
      "description": "Must-not-drop identity features.",
      "items": {
       "type": "string"
      },
      "title": "Signature",
      "type": "array"
     },
     "tag_description": {
      "items": {
       "type": "string"
      },
      "title": "Tag Description",
      "type": "array"
     },
     "trigger": {
      "default": "",
      "title": "Trigger",
      "type": "string"
     }
    },
    "required": [
     "id",
     "name",
     "gender",
     "age",
     "appearance",
     "description",
     "tag_description",
     "signature",
     "references",
     "outfits",
     "loras",
     "trigger"
    ],
    "title": "Character",
    "type": "object"
   },
   "CloneRequest": {
    "properties": {
     "title": {
      "default": "",
      "maxLength": 120,
      "title": "Title",
      "type": "string"
     }
    },
    "title": "CloneRequest",
    "type": "object"
   },
...
   "Episode": {
    "additionalProperties": false,
    "properties": {
     "base_prompt": {
      "default": "",
      "description": "Starting template (起手模板) copied into the prompt of new blank panels; never resolved at render time.",
      "title": "Base Prompt",
      "type": "string"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "deleted_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Deleted At"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "order": {
      "minimum": 0.0,
      "title": "Order",
      "type": "integer"
     },
     "panels": {
      "items": {
       "$ref": "#/components/schemas/Panel"
      },
      "title": "Panels",
      "type": "array"
     },
     "revision": {
      "default": 0,
      "title": "Revision",
      "type": "integer"
     },
     "series_id": {
      "title": "Series Id",
      "type": "string"
     },
     "strip": {
      "$ref": "#/components/schemas/Strip-Output"
     },
     "synopsis": {
      "default": "",
      "title": "Synopsis",
      "type": "string"
     },
     "takes": {
      "items": {
       "$ref": "#/components/schemas/Take"
      },
      "title": "Takes",
      "type": "array"
     },
     "title": {
      "title": "Title",
      "type": "string"
     },
     "updated_at": {
      "title": "Updated At",
      "type": "string"
     }
    },
    "required": [
     "id",
     "series_id",
     "title",
     "order",
     "synopsis",
     "base_prompt",
     "panels",
     "takes",
     "strip",
     "created_at",
     "updated_at",
     "deleted_at",
     "revision"
    ],
    "title": "Episode",
    "type": "object"
   },
...
   "Gender": {
    "enum": [
     "female",
     "male",
     "nonbinary",
     "unknown"
    ],
    "title": "Gender",
    "type": "string"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "Location-Output": {
    "additionalProperties": false,
    "properties": {
     "description": {
      "default": "",
      "title": "Description",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "references": {
      "items": {
       "$ref": "#/components/schemas/AssetRef-Output"
      },
      "title": "References",
      "type": "array"
     },
     "tags": {
      "items": {
       "type": "string"
      },
      "title": "Tags",
      "type": "array"
     },
     "time": {
      "default": "",
      "enum": [
       "",
       "morning",
       "day",
       "evening",
       "night"
      ],
      "title": "Time",
      "type": "string"
     }
    },
    "required": [
     "id",
     "name",
     "description",
     "tags",
     "time",
     "references"
    ],
    "title": "Location",
    "type": "object"
   },
...
   "Preset-Output": {
    "additionalProperties": false,
    "description": "A reusable visual asset: a named set of ``{变量}`` values (character, outfit, style…).\n\nPresets live in the workshop; assembling copies their values into the new album, so later\nedits never change an album that was already assembled.",
    "properties": {
     "entries": {
      "items": {
       "$ref": "#/components/schemas/PresetEntry-Output"
      },
      "title": "Entries",
      "type": "array"
     },
     "groups": {
      "items": {
       "$ref": "#/components/schemas/PresetGroup-Output"
      },
      "title": "Groups",
      "type": "array"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "title": {
      "title": "Title",
      "type": "string"
     }
    },
    "required": [
     "id",
     "title",
     "groups",
     "entries"
    ],
    "title": "Preset",
    "type": "object"
   },
...
   "PresetEntry-Output": {
    "additionalProperties": false,
    "description": "One ``{key}`` variable of a preset (legacy 预设工坊).",
    "properties": {
     "group_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Group Id"
     },
     "hint": {
      "default": "",
      "title": "Hint",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "key": {
      "pattern": "^[^\\s{}]{1,64}$",
      "title": "Key",
      "type": "string"
     },
     "label": {
      "default": "",
      "title": "Label",
      "type": "string"
     },
     "value": {
      "default": "",
      "title": "Value",
      "type": "string"
     }
    },
    "required": [
     "id",
     "key",
     "label",
     "value",
     "hint",
     "group_id"
    ],
    "title": "PresetEntry",
    "type": "object"
   },
...
   "PresetGroup-Output": {
    "additionalProperties": false,
    "properties": {
     "id": {
      "title": "Id",
      "type": "string"
     },
     "title": {
      "title": "Title",
      "type": "string"
     }
    },
    "required": [
     "id",
     "title"
    ],
    "title": "PresetGroup",
    "type": "object"
   },
   "PreviewRequest": {
    "properties": {
     "preset_ids": {
      "items": {
       "type": "string"
      },
      "minItems": 1,
      "title": "Preset Ids",
      "type": "array"
     },
     "profile_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Profile Id"
     },
     "prompt": {
      "maxLength": 8000,
      "minLength": 1,
      "title": "Prompt",
      "type": "string"
     },
     "title": {
      "default": "",
      "maxLength": 120,
      "title": "Title",
      "type": "string"
     }
    },
    "required": [
     "preset_ids",
     "prompt"
    ],
    "title": "PreviewRequest",
    "type": "object"
   },
...
   "Prop-Output": {
    "additionalProperties": false,
    "properties": {
     "description": {
      "default": "",
      "title": "Description",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "references": {
      "items": {
       "$ref": "#/components/schemas/AssetRef-Output"
      },
      "title": "References",
      "type": "array"
     },
     "tags": {
      "items": {
       "type": "string"
      },
      "title": "Tags",
      "type": "array"
     }
    },
    "required": [
     "id",
     "name",
     "description",
     "tags",
     "references"
    ],
    "title": "Prop",
    "type": "object"
   },
...
   "QueueAlbum": {
    "properties": {
     "album_id": {
      "title": "Album Id",
      "type": "string"
     }
    },
    "required": [
     "album_id"
    ],
    "title": "QueueAlbum",
    "type": "object"
   },
   "QueueAlbums": {
    "properties": {
     "album_ids": {
      "items": {
       "type": "string"
      },
      "maxItems": 2000,
      "title": "Album Ids",
      "type": "array"
     }
    },
    "title": "QueueAlbums",
    "type": "object"
   },
   "QueueSettings": {
    "properties": {
     "concurrency": {
      "anyOf": [
       {
        "maximum": 128.0,
        "minimum": 1.0,
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Concurrency"
     }
    },
    "title": "QueueSettings",
    "type": "object"
   },
   "QueueStatus": {
    "additionalProperties": false,
    "description": "The queue state plus what the page needs to explain it.",
    "properties": {
     "active": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "description": "Album the lane started last.",
      "title": "Active"
     },
     "active_job": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Active Job"
     },
     "auto_concurrency": {
      "description": "Panels per album when no default is set.",
      "title": "Auto Concurrency",
      "type": "integer"
     },
     "concurrency": {
      "anyOf": [
       {
        "maximum": 128.0,
        "minimum": 1.0,
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "description": "Default panels per album; None = automatic.",
      "title": "Concurrency"
     },
     "fault": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "description": "Storage warning (low disk space).",
      "title": "Fault"
     },
     "id": {
      "default": "queue",
      "title": "Id",
      "type": "string"
     },
     "lane": {
      "description": "Albums waiting to start one after another.",
      "items": {
       "type": "string"
      },
      "title": "Lane",
      "type": "array"
     },
     "notices": {
      "additionalProperties": {
       "type": "string"
      },
      "description": "Album id → why the lane could not start it.",
      "title": "Notices",
      "type": "object"
     },
     "order": {
      "description": "Album ids in card order; unlisted albums follow.",
      "items": {
       "type": "string"
      },
      "title": "Order",
      "type": "array"
     },
     "paused": {
      "default": false,
      "description": "The lane is held (全局暂停).",
      "title": "Paused",
      "type": "boolean"
     },
     "updated_at": {
      "title": "Updated At",
      "type": "string"
     }
    },
    "required": [
     "id",
     "order",
     "lane",
     "paused",
     "active",
     "active_job",
     "concurrency",
     "notices",
     "updated_at",
     "fault",
     "auto_concurrency"
    ],
    "title": "QueueStatus",
    "type": "object"
   },
...
   "Series": {
    "additionalProperties": false,
    "properties": {
     "bible": {
      "$ref": "#/components/schemas/Bible-Output"
     },
     "concurrency": {
      "anyOf": [
       {
        "maximum": 128.0,
        "minimum": 1.0,
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "description": "Panels this album renders at once (装配队列「并发」); None = the queue default.  A change applies to the running job from its next panel.",
      "title": "Concurrency"
     },
     "cover_asset_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "description": "Chosen cover image; None = automatic (the first image).",
      "title": "Cover Asset Id"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "default_profile_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Default Profile Id"
     },
     "deleted_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Deleted At"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "kind": {
      "default": "album",
      "description": "album = a book on the shelf; workshop = the storyboards and presets of the 创作工坊 (never listed as a book).",
      "enum": [
       "album",
       "workshop"
      ],
      "title": "Kind",
      "type": "string"
     },
     "presets": {
      "description": "Workshop presets.",
      "items": {
       "$ref": "#/components/schemas/Preset-Output"
      },
      "title": "Presets",
      "type": "array"
     },
     "status": {
      "$ref": "#/components/schemas/SeriesStatus",
      "default": "draft"
     },
     "subtitle": {
      "default": "",
      "title": "Subtitle",
      "type": "string"
     },
     "title": {
      "title": "Title",
      "type": "string"
     },
     "updated_at": {
      "title": "Updated At",
      "type": "string"
     },
     "variables": {
      "additionalProperties": {
       "type": "string"
      },
      "description": "{变量} values (advanced layer).",
      "title": "Variables",
      "type": "object"
     },
     "variants": {
      "items": {
       "$ref": "#/components/schemas/VariantSet-Output"
      },
      "title": "Variants",
      "type": "array"
     }
    },
    "required": [
     "id",
     "title",
     "subtitle",
     "status",
     "kind",
     "presets",
     "bible",
     "variants",
     "variables",
     "default_profile_id",
     "cover_asset_id",
     "concurrency",
     "created_at",
     "updated_at",
     "deleted_at"
    ],
    "title": "Series",
    "type": "object"
   },
...
   "SeriesStatus": {
    "enum": [
     "draft",
     "active",
     "archived"
    ],
    "title": "SeriesStatus",
    "type": "string"
   },
...
   "Style-Output": {
    "additionalProperties": false,
    "properties": {
     "description": {
      "default": "",
      "title": "Description",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "loras": {
      "items": {
       "$ref": "#/components/schemas/LoraRef-Output"
      },
      "title": "Loras",
      "type": "array"
     },
     "name": {
      "default": "默认画风",
      "title": "Name",
      "type": "string"
     },
     "negative": {
      "items": {
       "type": "string"
      },
      "title": "Negative",
      "type": "array"
     },
     "references": {
      "items": {
       "$ref": "#/components/schemas/AssetRef-Output"
      },
      "title": "References",
      "type": "array"
     },
     "tag_description": {
      "items": {
       "type": "string"
      },
      "title": "Tag Description",
      "type": "array"
     }
    },
    "required": [
     "id",
     "name",
     "description",
     "tag_description",
     "negative",
     "references",
     "loras"
    ],
    "title": "Style",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
   "VariantSet-Output": {
    "additionalProperties": false,
    "description": "Batch variant (the old \"character matrix\"): the same script rendered with another bible set.\n\n``characters`` maps a character id to a partial Character patch (appearance, tags, loras…).",
    "properties": {
     "characters": {
      "additionalProperties": {
       "additionalProperties": true,
       "type": "object"
      },
      "title": "Characters",
      "type": "object"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "profile_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Profile Id"
     },
     "style_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Style Id"
     }
    },
    "required": [
     "id",
     "name",
     "characters",
     "style_id",
     "profile_id"
    ],
    "title": "VariantSet",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "bearer": {
    "bearerFormat": "mio_<id>_<secret>",
    "scheme": "bearer",
    "type": "http"
   }
  }
 },
 "info": {
  "description": "Open API of Mio Comic Studio. Authenticate with `Authorization: Bearer mio_…` (create tokens in Settings → API). Scopes: read, write, render, admin.",
  "title": "Mio API",
  "version": "2.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/workshop": {
   "get": {
    "description": "The hidden workshop series: its episodes are the storyboards, ``presets`` the presets.\nCreated (seeded from the legacy ``data/``) on first use.",
    "operationId": "get_workshop",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Series"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Get Workshop",
    "tags": [
     "workshop"
    ]
   }
  },
  "/workshop/assemble": {
   "post": {
    "description": "Storyboard + presets → a new album waiting in the queue.  Makes no model call; start it\nwith ``POST /episodes/{id}/render``.",
    "operationId": "assemble",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AssembleRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Assembled"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Assemble",
    "tags": [
     "workshop"
    ]
   }
  },
  "/workshop/preview": {
   "post": {
    "description": "预设工坊「独立试绘」: one prompt + presets → a one-frame album waiting in the queue (no\nstoryboard, no model call).  Start it with ``POST /episodes/{id}/render``.",
    "operationId": "preview_preset",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PreviewRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Assembled"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Preview Preset",
    "tags": [
     "workshop"
    ]
   }
  },
  "/workshop/queue": {
   "get": {
    "description": "装配队列: card order, the sequential lane, the default concurrency and storage warnings.",
    "operationId": "get_queue",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/QueueStatus"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Get Queue",
    "tags": [
     "workshop"
    ]
   },
   "patch": {
    "description": "默认并发: panels per album for albums without their own setting (None = automatic).",
    "operationId": "patch_queue",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/QueueSettings"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/QueueStatus"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Patch Queue",
    "tags": [
     "workshop"
    ]
   }
  },
  "/workshop/queue/clear": {
   "post": {
    "description": "停止全部: empty the lane (running jobs are stopped through the job API).",
    "operationId": "clear_queue",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/QueueStatus"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Clear Queue",
    "tags": [
     "workshop"
    ]
   }
  },
  "/workshop/queue/order": {
   "put": {
    "description": "Drag to reorder the task cards; the lane follows the new order.",
    "operationId": "order_queue",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/QueueAlbums"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/QueueStatus"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Order Queue",
    "tags": [
     "workshop"
    ]
   }
  },
  "/workshop/queue/pause": {
   "post": {
    "description": "Hold the lane (全局暂停 also pauses the running jobs, through the job API).",
    "operationId": "pause_queue",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/QueueStatus"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Pause Queue",
    "tags": [
     "workshop"
    ]
   }
  },
  "/workshop/queue/remove": {
   "post": {
    "description": "移出队列: the album stops waiting in the lane; jobs already started are not touched.",
    "operationId": "remove_from_queue",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/QueueAlbum"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/QueueStatus"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Remove From Queue",
    "tags": [
     "workshop"
    ]
   }
  },
  "/workshop/queue/resume": {
   "post": {
    "operationId": "resume_queue",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/QueueStatus"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Resume Queue",
    "tags": [
     "workshop"
    ]
   }
  },
  "/workshop/queue/start": {
   "post": {
    "description": "按顺序开始生成: the albums start one after another (each renders its missing panels).",
    "operationId": "start_queue",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/QueueAlbums"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/QueueStatus"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Start Queue",
    "tags": [
     "workshop"
    ]
   }
  },
  "/workshop/tasks/{series_id}/clone": {
   "post": {
    "description": "装配队列「克隆」: a new standby album with the same frames, presets and profile (no images).",
    "operationId": "clone_task",
    "parameters": [
     {
      "in": "path",
      "name": "series_id",
      "required": true,
      "schema": {
       "title": "Series Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/CloneRequest"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Assembled"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "bearer": []
     }
    ],
    "summary": "Clone Task",
    "tags": [
     "workshop"
    ]
   }
  }
 },
 "servers": [
  {
   "url": "/api/v2"
  }
 ]
}